+++ wfe-web/src/main/webapp/WEB-INF/af/chat_view.jsp
@@ -7,7 +7,7 @@
     <meta http-equiv="content-type" content="text/html; charset=UTF-8">
     <link rel="stylesheet" type="text/css" href="<html:rewrite page='<%="/css/chat.css?"+Version.getHash() %>' />">
     <script type="text/javascript" src="<html:rewrite page="/js/jquery-1.8.3.min.js" />">c = 0;</script>
-    <script type="text/javascript" src="<html:rewrite page='<%="/js/chatSending.js" %>' />">c = 0;</script>
+    <script type="text/javascript" src="<html:rewrite page='<%="/js/chatHandlers.js" %>' />">c = 0;</script>
 </head>
 <body>
 <script type="text/javascript">
